Crate eq_rate

Equilibrium Rate Pallet

Equilibrium's Rate Pallet is a Substrate module for processing different fee payments and keeping account balances up to date

Re-exports

pub use eq_oracle;

Modules

ed25519
Module for crypto signatures

reinit_extension
A SignedExtension module. Adds a reinit request for the signer of each trx

Structs

GenesisConfig
Genesis config for the module, allow to build genesis storage.

LastFeeUpdate
Pallet storage - last update timestamps for each AccountId that has balances

Module

NowMillisOffset
Pallet storage used for time offset in test builds

ReinitRequest
Transfers service data inside Rate Pallet functions

Enums

Call
Dispatchable calls.

Error
Pallet's errors

Constants

KEY_TYPE

Traits

Trait
Substrate pallet configuration trait

Type Definitions

AuthIndex