Trait eq_rate::Trait

pub trait Trait: SendTransactionTypes<Call<Self>> + Trait {
    type UnixTime: UnixTime;
    type Balance: Member + AtLeast32Bit + MaybeSerializeDeserialize + Parameter + Default + From<u64> + Into<u64>;
    type BalanceGetter: BalanceGetter<Self::AccountId, Self::Balance>;
    type AuthorityId: Member + Parameter + RuntimeAppPublic + Default + Ord;
    type BailsmanManager: BailsmanManager<Self::AccountId>;
    type LtvChecker: LtvChecker<Self::AccountId, Self::Balance>;
    type MinSurplus: Get<Self::Balance>;
    type MinTempBailsman: Get<Self::Balance>;
    type PriceGetter: PriceGetter;
    type UnsignedPriority: Get<TransactionPriority>;
    type FeeManager: FeeManager<Self::AccountId, Self::Balance>;
    type EqBuyout: EqBuyout<Self::AccountId, Self::Balance>;
    type BailsmanModuleId: Get<ModuleId>;
}

Substrate pallet configuration trait

Associated Types

type UnixTime: UnixTime

Timestamp provider

type Balance: Member + AtLeast32Bit + MaybeSerializeDeserialize + Parameter + Default + From<u64> + Into<u64>

Numerical representation of stored balances

type BalanceGetter: BalanceGetter<Self::AccountId, Self::Balance>

Gets information about account balances

type AuthorityId: Member + Parameter + RuntimeAppPublic + Default + Ord

The identifier type for an authority.

type BailsmanManager: BailsmanManager<Self::AccountId>

Used to integrate bailsman operations

type LtvChecker: LtvChecker<Self::AccountId, Self::Balance>

Calculates and checks LTV of user balances

type MinSurplus: Get<Self::Balance>

Minimum new debt for system reinit

type MinTempBailsman: Get<Self::Balance>

Minimum temp bailsmen balances for Bailsman pallet reinit

type PriceGetter: PriceGetter

Receives currency prices from oracle

type UnsignedPriority: Get<TransactionPriority>

A configuration for base priority of unsigned transactions.

This is exposed so that it can be tuned for particular runtime, when multiple pallets send unsigned transactions.

type FeeManager: FeeManager<Self::AccountId, Self::Balance>

Manages all fee calculations and withdrawal

type EqBuyout: EqBuyout<Self::AccountId, Self::Balance>

Manager for treasury Eq exchanging transactions

type BailsmanModuleId: Get<ModuleId>

Gets bailsman module account

Loading content...